+++ src/CustomType/Data/Field/ColorField.php
@@ -2,6 +2,8 @@
 
 namespace Torr\PrismicApi\CustomType\Data\Field;
 
+use Torr\PrismicApi\CustomType\Helper\FilterFieldsHelper;
+
 /**
  * @see https://prismic.io/docs/core-concepts/color
  */
@@ -17,7 +19,7 @@ final class ColorField extends InputField
 		string $label,
 	)
 	{
-		parent::__construct(self::TYPE_KEY, $this->filterOptionalFields([
+		parent::__construct(self::TYPE_KEY, FilterFieldsHelper::filterOptionalFields([
 			"label" => $label,
 		]));
 	}