+++ src/CustomType/Data/Field/InputField.php
@@ -29,14 +29,4 @@ class InputField implements PrismicTypeInterface
 			"config" => $this->config,
 		];
 	}
-
-	/**
-	 */
-	protected function filterOptionalFields (array $config) : array
-	{
-		return \array_filter(
-			$config,
-			static fn ($entry) => null !== $entry,
-		);
-	}
 }